Renamed ImageLink
renamed ImageLink to PictureCard
restyled pictureCard
added dark layer to image background
