--- src/components/ImageLink/imageLink.test.jsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import { shallow } from 'enzyme';
-import ImageLink from './index';
-import cardImage from './cards_bilde.jpg';
-
-describe('ImageLink component', () => {
-	test('Matches the snapshot', () => {
-		const wrapper = shallow(<ImageLink image={<img src={cardImage} alt="image" />} onClick={() => {}} type="Type" title="Tittel" />);
-		expect(wrapper).toMatchSnapshot();
-	});
-	test('Sets default orientation className', () => {
-		const wrapper = shallow(<ImageLink image={<img src={cardImage} alt="image" />} onClick={() => {}} type="Type" title="Tittel" />);
-		expect(wrapper.hasClass('vertical')).toEqual(true);
-	});
-	test('Switches className on orientation change', () => {
-		const wrapper = shallow(<ImageLink orientation="horizontal" image={<img src={cardImage} alt="image" />} onClick={() => {}} type="Type" title="Tittel" />);
-		expect(wrapper.hasClass('horizontal')).toEqual(true);
-	});
-});
