+++ src/main.scss
@@ -22,7 +22,7 @@
 @import './components/FormError/form-error';
 @import './components/Glossary/glossary';
 @import './components/Header/header';
-@import './components/ImageLink/image-link';
+@import 'src/components/PictureCard/picture-card';
 @import './components/Input/input';
 @import './components/InputError/input-error';
 @import './components/KeyFigures/key-figures';
